Fix incorrect narrowing of a factory input value left over from an older version of the test case reproduction interpreter - the input values are now `BigInt` and should be left as such. Fixes outstanding test failure.

+++ src/main/scala/com/sageserpent/americium/TrialsImplementation.scala
@@ -88,7 +88,7 @@ case class TrialsImplementation[Case](
                 ) =
                   decisionStages: @unchecked
                 _ <- State.set(Context(remainingDecisionStages, 1 + complexity))
-              } yield factory(input.toInt)
+              } yield factory(input)
 
             // NOTE: pattern-match only on `Some`, as we are reproducing a case
             // that by dint of being reproduced, must have passed filtration the
